[Refactor] 댓글 조회 api 분리
